Restructure application service a bit
