--- foo/foo-dal.test.js
@@ -1,7 +1,0 @@
-'use strict';
-
-// const createFooDAL = require('./foo-dal');
-
-describe('foo-dal', () => {
-    it.todo('should get a foo');
-});
