--- foo/foo-routes.test.js
@@ -1,5 +1,0 @@
-'use strict';
-
-describe('foo-routes', () => {
-    it.todo('should route correctly');
-});
