--- foo/foo-service.test.js
@@ -1,7 +1,0 @@
-'use strict';
-
-// const createFooService = require('./foo-service');
-
-describe('foo-service', () => {
-    it.todo('should get a foo resource');
-});
