+++ kube_deploy/Dev/ingress.yml
@@ -3,9 +3,9 @@ kind: Ingress
 metadata:
   annotations:
     nginx.ingress.kubernetes.io/whitelist-source-range: 31.121.104.9/32,37.143.143.162/32,37.143.143.10/32
-    external-dns.alpha.kubernetes.io/set-identifier: app-service-ingress-dev-claim-criminal-injuries-compensation-dev-green
+    external-dns.alpha.kubernetes.io/set-identifier: application-service-ingress-dev-claim-criminal-injuries-compensation-dev-green
     external-dns.alpha.kubernetes.io/aws-weight: '100'
-  name: app-service-ingress-dev
+  name: application-service-ingress-dev
   namespace: claim-criminal-injuries-compensation-dev
 spec:
   tls:
@@ -22,7 +22,7 @@ spec:
           - path: /
             backend:
               service:
-                name: app-service-dev
+                name: application-service-dev
                 port:
                   number: 80
             pathType: ImplementationSpecific
@@ -32,7 +32,7 @@ spec:
           - path: /
             backend:
               service:
-                name: app-service-dev
+                name: application-service-dev
                 port:
                   number: 80
             pathType: ImplementationSpecific
