+++ services/sqs/index.test.js
@@ -1,0 +1,15 @@
+'use strict';
+
+// const fs = require('fs');
+// const createSqsService = require('./index');
+
+// describe('SQS Service', () => {
+//     it('Should read SQS message in correct format', async () => {
+//         const path = 'resources/testing/output.pdf';
+//         const stream = fs.readFileSync('resources/testing/checkYourAnswers.json');
+//         const sqsService = createPdfService();
+//         await pdfService.writeJSONToPDF(JSON.parse(stream), path);
+//         expect(fs.existsSync(path)).toBeTruthy();
+//     });
+// });
+
